ml/dl/examples/handwritten_with_gradio/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.13
altair==5.0.1
altgraph==0.17.3
annotated-types==0.5.0
anyio==3.7.1
astroid==2.15.5
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
Babel==2.12.1
backports.tempfile==1.0
backports.weakref==1.0.post1
bleach==6.0.0
bump2version==1.0.1
bumpversion==0.6.0
cachetools==5.3.1
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.7
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
conan==2.0.9
constantly==15.1.0
contourpy==1.1.0
coverage==7.2.7
cycler==0.11.0
decorator==5.1.1
dill==0.3.6
distlib==0.3.6
docutils==0.18.1
exceptiongroup==1.1.1
fastapi==0.101.1
fasteners==0.18
ffmpy==0.3.1
filelock==3.12.0
flake8==4.0.1
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.3.3
fsspec==2023.6.0
gradio==3.40.1
gradio_client==0.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
humanize==4.6.0
hyperlink==21.0.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.6.0
importlib-resources==6.0.1
incremental==22.10.0
iniconfig==2.0.0
isort==5.12.0
jaraco.classes==3.2.3
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==23.13.1
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
linkify-it-py==2.0.2
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mccabe==0.6.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.2.1
multidict==6.0.4
mypy==1.3.0
mypy-extensions==1.0.0
networkx==3.0
node==1.2.1
numpy==1.24.1
odict==1.9.0
onnx==1.14.0
onnxruntime==1.15.1
opencv-python==4.8.0.76
orjson==3.9.5
packaging==23.1
pandas==2.0.3
passlib==1.7.4
patch-ng==1.17.4
pefile==2023.2.7
pep8-naming==0.13.2
Pillow==9.5.0
pkginfo==1.9.6
platformdirs==3.5.1
plumber==1.7
protobuf==4.24.1
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pydantic==2.2.1
pydantic_core==2.6.1
pydub==0.25.1
pyenchant==3.2.2
pyflakes==2.4.0
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pylint==2.17.4
pyparsing==3.0.9
pyproject_api==1.5.1
pyreadline3==3.4.1
pyte==0.8.1
pytest==7.3.1
pytest-aiohttp==1.0.4
pytest-asyncio==0.21.0
pytest-runner==6.0.0
python-dateutil==2.8.2
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
pywin32-ctypes==0.2.0
PyYAML==6.0
qualname==0.1.0
readme-renderer==37.3
referencing==0.30.2
requests==2.28.1
requests-toolbelt==1.0.0
rpds-py==0.9.2
semantic-version==2.10.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==6.2.1
sphinx-autoapi==2.1.0
sphinx-rtd-theme==1.2.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-images==0.9.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
starlette==0.27.0
sympy==1.11.1
thefuck==3.32
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tox==4.6.0
tox-gh-actions==3.1.1
tqdm==4.66.1
twine==4.0.2
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
Unidecode==1.3.6
urllib3==1.26.13
uvicorn==0.23.2
virtualenv==20.23.0
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3
win-unicode-console==0.5
wrapt==1.15.0
XlsxWriter==3.1.2
yapf==0.29.0
yarl==1.9.2
zipp==3.15.0
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==5.4
zope.interface==6.0
zope.lifecycleevent==5.0
zope.proxy==5.0.0
